fix: address Claude review — move lazy TTS to live endpoint, drop dead code
Claude Code Review on PR #710 correctly identified that backend/routers/teachers.py
is shadowed by the backend/routers/teachers/ package. Production resolves
`from routers import teachers` to the package's __init__.py; the file is
never imported. The lazy TTS pre-pass added in this PR's first commit lived
at teachers.py:3878-3886 — dead code.

Confirmed by curl on staging: GET /api/teachers/assignments/181/preview/practice-words
returns 404 (route not registered). PR #709's preview_practice_words endpoint
also lived in the dead file and never deployed.

Changes:

1. routers/teachers/assignment_ops.py — the LIVE module:
   - Add lazy TTS pre-pass to get_assignment_preview() before
     build_assignment_preview() call. Mirrors the pattern in
     routers/students/assignments.get_assignment_activities. Mode-gated
     loads (only for vocab content types), idempotent helpers
     (ensure_example_sentence_audio / ensure_word_audio).
   - Add the preview_practice_words endpoint that PR #709 intended to
     ship — uses the existing _get_teacher_assignment helper and
     returns session_id=None, no PracticeSession row.
   - Module-level imports (not inline-conditional) per reviewer's
     style note.

2. routers/teachers.py — DEAD file:
   - Remove the lazy TTS pre-pass (this commit) and the
     preview_practice_words endpoint (from PR #709) that were never
     reachable in production.
   - Restore the original `from services.preview_service import
     get_sentence_fields` (without the dead _VOCABULARY_CONTENT_TYPES
     re-import).

3. services/preview_service.py — unchanged from first commit
   (`WHERE col IS NULL OR col = ''` fix is correct and applies live).

Verified live router registration:
    practice-words routes: ['/assignments/{assignment_id}/preview/practice-words']
    main preview route: ['/assignments/{assignment_id}/preview']

Out of scope per reviewer's "minor #3" note: the audio_by_id return value
from the lazy TTS helpers is still discarded (matching student endpoint
pattern). If the SQLAlchemy expire-on-commit silent-failure case ever
manifests, fix at all three call sites in one pass — not in this PR.
